Move events.ts out of lang and rename Event to MalloyEvent to prevent shadowing built in Event
