+++ packages/malloy/src/lang/parse-malloy.ts
@@ -90,7 +90,7 @@ import {Tag} from '../tags';
 import {MalloyParseInfo} from './malloy-parse-info';
 import {walkForModelAnnotation} from './parse-tree-walkers/model-annotation-walker';
 import {walkForTablePath} from './parse-tree-walkers/find-table-path-walker';
-import {EventStream} from './events';
+import {EventStream} from '../events';
 
 export type StepResponses =
   | DataRequestResponse
